Add authorization to organization controller

+++ backend/apid/controllers/organizations.go
@@ -7,18 +7,22 @@ import (
 	"net/http"
 
 	"github.com/gorilla/mux"
+	"github.com/sensu/sensu-go/backend/authorization"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/types"
 )
 
 // OrganizationsController defines the fields required for this controller.
 type OrganizationsController struct {
-	Store store.Store
+	Store     store.Store
+	abilities authorization.Ability
 }
 
 // Register should define an association between HTTP routes and their
 // respective handlers defined within this Controller.
 func (o *OrganizationsController) Register(r *mux.Router) {
+	o.abilities = authorization.Ability{Resource: types.RuleTypeOrganization}
+
 	r.HandleFunc("/rbac/organizations", o.many).Methods(http.MethodGet)
 	r.HandleFunc("/rbac/organizations", o.update).Methods(http.MethodPost)
 	r.HandleFunc("/rbac/organizations/{organization}", o.single).Methods(http.MethodGet)
@@ -29,6 +33,14 @@ func (o *OrganizationsController) Register(r *mux.Router) {
 func (o *OrganizationsController) delete(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	org := vars["organization"]
+
+	abilities := o.abilities.WithContext(r.Context())
+	abilities.Actor.Organization = org
+
+	if !abilities.CanDelete() {
+		authorization.UnauthorizedAccessToResource(w)
+		return
+	}
 
 	err := o.Store.DeleteOrganizationByName(r.Context(), org)
 	if err != nil {
@@ -41,6 +53,12 @@ func (o *OrganizationsController) delete(w http.ResponseWriter, r *http.Request)
 
 // many returns all organizations
 func (o *OrganizationsController) many(w http.ResponseWriter, r *http.Request) {
+	abilities := o.abilities.WithContext(r.Context())
+	if !abilities.CanRead() {
+		authorization.UnauthorizedAccessToResource(w)
+		return
+	}
+
 	orgs, err := o.Store.GetOrganizations(r.Context())
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusInternalServerError)
@@ -60,6 +78,14 @@ func (o *OrganizationsController) many(w http.ResponseWriter, r *http.Request) {
 func (o *OrganizationsController) single(w http.ResponseWriter, r *http.Request) {
 	vars := mux.Vars(r)
 	name := vars["organization"]
+
+	abilities := o.abilities.WithContext(r.Context())
+	abilities.Actor.Organization = name
+
+	if !abilities.CanRead() {
+		authorization.UnauthorizedAccessToResource(w)
+		return
+	}
 
 	var (
 		org *types.Organization
@@ -98,6 +124,14 @@ func (o *OrganizationsController) update(w http.ResponseWriter, r *http.Request)
 		return
 	}
 	defer r.Body.Close()
+
+	abilities := o.abilities.WithContext(r.Context())
+	abilities.Actor.Organization = org.Name
+
+	if !abilities.CanCreate() {
+		authorization.UnauthorizedAccessToResource(w)
+		return
+	}
 
 	err = json.Unmarshal(bodyBytes, &org)
 	if err != nil {
